Remove test directory and files
The Kong microservice has no custom code, and is simply running the Kong
application and syncing configuration that is tracked in the config
file. Outside of the config file and the cloud.gov files needed to
deploy the application, there is nothing we can meaningfully unit-test.
Due to this, we are removing the test files, and will continue to run
integration tests as normal.