Move Kong configuration to kong.conf file
Removing all Kong configuration from the manifest.yaml file helps keep
all of the Kong configuration in the one place it's supposed to be.
There is now a single point of truth for Kong's configuration, as
opposed to relying on environment variable precedent and potentially
overriding the kong.conf file unnecessarily or by accident.