Move Kong to a closed-egress space
Kong will have a public route attached to it. It will be able to respond
to incoming traffic, but does not need to initiate any connections to
the public internet or internal brokered services.

The route from Kong to sk-api now must specify the space that the sk-api
will be in, which should be the default environment space (dev, test,
prod).