Reduce Kong memory allocation (#132)
After reducing the number of Kong worker processes from the auto setting
of $num_cpu_cores to 1, we've seen the expected reduction in memory
usage we were looking for and can safely reduce the amount of memory
allocated to the Kong app. CPU utilization remains incredibly low, so
reducing the allocated memory will be a big improvement of resource
utilization.