Use CF Instance Identity Certs for status API
In order to use fully-validated TLS connections to the Kong status API,
we can configure the status API to utilize the CF Instance Identity
certificates, which contain the IP SAN for the application. This allows
Prometheus to connect to Kong and validate the certificate as normal,
since Prometheus is using the IP address, rather than the route, to
connect to Kong for metric info.